pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]
exclude = [
    "/migrations/",
]

[tool.poetry]
name = "match_prediction_app"
version = "0.1.0"
description = ""
authors = [
    "Sofiane Bebert <s.bebert@diceytech.co.uk>",
]
license = ""
readme = "README.md"
packages = [
    { include = "match_prediction_app" },
]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.5.0"
argon2-cffi = "^21.3.0"
hiredis = "^2.2.3"
redis = "^4.6.0"
Django = "^4.2.2"
django-environ = "^0.10.0"
django-model-utils = "^4.3.1"
django-allauth = "^0.54.0"
django-render-block = "^0.9.2"
django-extensions = "^3.2.3"
django-improved-user = "2.0a2"
django-htmx = "^1.15.0"
django-anymail = "^10.0"
boto3 = "^1.26.160"
gunicorn = "^20.1.0"
sentry-sdk = "^1.26.0"
fuzzy-couscous = "^3.1.1"
pre-commit = "^3.4.0"

[tool.poetry.dependencies.whitenoise]
extras = [
    "brotli",
]
version = "^6.5.0"

[tool.poetry.group.dev.dependencies]
django-browser-reload = "^1.10.0"
django-debug-toolbar = "^4.1.0"
django-fastdev = "^1.7.3"
ipython = "^8.14.0"
pywatchman = "^1.4.1"
coverage = "^7.2.7"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.7"
mypy = "^1.4.1"
django-stubs = "^4.2.1"
bandit = "^1.7.5"
poethepoet = "^0.20.0"

[tool.poetry.group.dev.dependencies.Werkzeug]
extras = [
    "watchdog",
]
version = "^2.3.6"

[tool.poe.tasks]
RD = [
    "rd",
    "rm",
    "mk",
    "m",
]

[tool.poe.tasks.r]
cmd = "python manage.py runserver"
help = "Start dev serve"

[tool.poe.tasks.u]
cmd = "python manage.py makesuperuser"
help = "Create a new superuser using environment varibales, look in your .env.template file"

[tool.poe.tasks.m]
cmd = "python manage.py migrate"
help = "Apply migrations"

[tool.poe.tasks.mk]
cmd = "python manage.py makemigrations"
help = "Make migrations"

[tool.poe.tasks.rd]
cmd = "python manage.py reset_db --noinput"
help = "Delete and recreate your database"

[tool.poe.tasks.t]
cmd = "tailwindcss -i match_prediction_app/static/css/input.css -o match_prediction_app/static/css/output.css --watch"
help = "Compile tailwind in watch mode"

[tool.poe.tasks.rm]
cmd = "cuzzy rm-migrations match_prediction_app"
help = "Delete all migrations files"

[tool.cuzzy]
work = [
    "poe r",
    "poe t",
]

[tool.ruff]
line-length = 120
select = [
    "E",
    "F",
    "UP",
    "N",
    "S",
    "C",
    "B",
    "T",
]
extend-exclude = [
    "*/migrations/*",
]

[tool.django-stubs]
django_settings_module = "config.settings"
